Use correct type for g_base64_decode to fix stack corruption on x86_64 (#522389).
svn path=/trunk/; revision=8586

+++ servers/groupwise/e-gw-connection.c
@@ -2694,7 +2694,7 @@ e_gw_connection_get_attachment (EGwConnection *cnc, const char *id, int offset, 
 	}
 
 	if (buffer && buf_length) {
-		int len = atoi (buf_length) ;
+		gsize len = atoi (buf_length) ;
 		*attachment = g_base64_decode (buffer,&len) ;
 		*attach_length = len ;
 	}
